Dreamview: audio event editor
