+++ modules/dreamview/frontend/proto_bundle/sim_world_proto_bundle.json
@@ -294,6 +294,17 @@
                 "obstaclePriority": {
                   "type": "apollo.prediction.ObstaclePriority",
                   "id": 33
+                },
+                "source": {
+                  "type": "apollo.perception.PerceptionObstacle.Source",
+                  "id": 36,
+                  "options": {
+                    "default": "HOST_VEHICLE"
+                  }
+                },
+                "v2xInfo": {
+                  "type": "apollo.perception.V2XInformation",
+                  "id": 37
                 }
               },
               "nested": {
@@ -3843,6 +3854,10 @@
                   "rule": "repeated",
                   "type": "string",
                   "id": 2
+                },
+                "localRouting": {
+                  "type": "RoutingResponseFeature",
+                  "id": 3
                 }
               }
             },
@@ -3980,8 +3995,8 @@
                   "type": "apollo.common.Header",
                   "id": 1
                 },
-                "learningData": {
-                  "type": "LearningData",
+                "learningDataFrame": {
+                  "type": "LearningDataFrame",
                   "id": 2
                 }
               }
@@ -6425,7 +6440,9 @@
                     "NO_LEARNING": 0,
                     "E2E": 1,
                     "HYBRID": 2,
-                    "RL_TEST": 3
+                    "RL_TEST": 3,
+                    "E2E_TEST": 4,
+                    "HYBRID_TEST": 5
                   }
                 }
               }
@@ -7371,7 +7388,8 @@
                   "values": {
                     "CONV_RNN": 1,
                     "CNN": 2,
-                    "CNN_LSTM": 3
+                    "SELF_CNN_LSTM": 3,
+                    "HISTORY_UNCONDITIONED_CNN_LSTM": 4
                   }
                 }
               }
@@ -7743,6 +7761,13 @@
                   "options": {
                     "default": 1.5
                   }
+                },
+                "adcBufferCoeff": {
+                  "type": "double",
+                  "id": 7,
+                  "options": {
+                    "default": 1
+                  }
                 }
               }
             },
@@ -7773,6 +7798,10 @@
                   "options": {
                     "default": 20
                   }
+                },
+                "weightXRefPathReference": {
+                  "type": "double",
+                  "id": 2
                 }
               }
             },
@@ -7874,6 +7903,13 @@
                 "laneChangePathConfig": {
                   "type": "PiecewiseJerkPathWeights",
                   "id": 2
+                },
+                "pathReferenceLWeight": {
+                  "type": "double",
+                  "id": 3,
+                  "options": {
+                    "default": 0
+                  }
                 }
               }
             },
@@ -9745,6 +9781,10 @@
                 "learningModelOutputFailReason": {
                   "type": "string",
                   "id": 3
+                },
+                "evaluatedPathReference": {
+                  "type": "apollo.common.Path",
+                  "id": 4
                 }
               }
             }
@@ -9973,6 +10013,27 @@
                 }
               }
             },
+            "V2XInformation": {
+              "fields": {
+                "v2xType": {
+                  "rule": "repeated",
+                  "type": "V2XType",
+                  "id": 1,
+                  "options": {
+                    "packed": false
+                  }
+                }
+              },
+              "nested": {
+                "V2XType": {
+                  "values": {
+                    "NONE": 0,
+                    "ZOMBIES_CAR": 1,
+                    "BLIND_ZONE": 2
+                  }
+                }
+              }
+            },
             "SensorMeasurement": {
               "fields": {
                 "sensorId": {
@@ -10171,6 +10232,17 @@
                 "msg": {
                   "type": "DebugMessage",
                   "id": 26
+                },
+                "source": {
+                  "type": "Source",
+                  "id": 27,
+                  "options": {
+                    "default": "HOST_VEHICLE"
+                  }
+                },
+                "v2xInfo": {
+                  "type": "V2XInformation",
+                  "id": 28
                 }
               },
               "nested": {
@@ -10205,6 +10277,12 @@
                     "ST_TRICYCLIST": 9,
                     "ST_PEDESTRIAN": 10,
                     "ST_TRAFFICCONE": 11
+                  }
+                },
+                "Source": {
+                  "values": {
+                    "HOST_VEHICLE": 0,
+                    "V2X": 1
                   }
                 }
               }
@@ -12614,6 +12692,76 @@
                 "navigationLane": {
                   "type": "NavigationLaneConfig",
                   "id": 2
+                }
+              }
+            }
+          }
+        },
+        "audio": {
+          "nested": {
+            "MovingResult": {
+              "values": {
+                "UNKNOWN": 0,
+                "APPROACHING": 1,
+                "DEPARTING": 2,
+                "STATIONARY": 3
+              }
+            },
+            "AudioType": {
+              "values": {
+                "UNKNOWN_TYPE": 0,
+                "POLICE": 1,
+                "AMBULANCE": 2,
+                "FIRETRUCK": 3
+              }
+            },
+            "AudioDirection": {
+              "values": {
+                "UNKNOWN_DIRECTION": 0,
+                "FRONT": 1,
+                "LEFT": 2,
+                "BACK": 3,
+                "RIGHT": 4
+              }
+            },
+            "AudioEvent": {
+              "fields": {
+                "header": {
+                  "type": "apollo.common.Header",
+                  "id": 1
+                },
+                "id": {
+                  "type": "int32",
+                  "id": 2
+                },
+                "movingResult": {
+                  "type": "apollo.audio.MovingResult",
+                  "id": 3,
+                  "options": {
+                    "default": "UNKNOWN"
+                  }
+                },
+                "audioType": {
+                  "type": "apollo.audio.AudioType",
+                  "id": 4,
+                  "options": {
+                    "default": "UNKNOWN_TYPE"
+                  }
+                },
+                "sirenIsOn": {
+                  "type": "bool",
+                  "id": 5
+                },
+                "audioDirection": {
+                  "type": "apollo.audio.AudioDirection",
+                  "id": 6,
+                  "options": {
+                    "default": "UNKNOWN_DIRECTION"
+                  }
+                },
+                "pose": {
+                  "type": "apollo.localization.Pose",
+                  "id": 7
                 }
               }
             }
