+++ modules/dreamview/frontend/src/components/SideBar/POI.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { observer } from 'mobx-react';
-import classNames from 'classnames';
 
 import RadioItem from 'components/common/RadioItem';
